+++ changelog.md
@@ -24,6 +24,7 @@
   is specified, it is used to overwrite corresponding information in the constructed node.
 - Generalized `cai_causal_graph.causal_graph.CausalGraph.__eq__` to check for the class of the instance itself, enabling
   to reuse this method by extending classes.
+- Added `cai_causal_graph.causal_graph.CausalGraph.has_non_serializable_metadata` method, which returns `False` by default.
 - Dropped support for `python` `3.8` as it is approaching end of life.
 
 ## 0.4.10